+++ src/lib.rs
@@ -691,6 +691,7 @@ pub mod shared;
 pub mod test_app;
 
 pub mod prelude {
+    #[expect(deprecated, reason = "Re-export of deprecated aliases")]
     pub use super::{
         RepliconPlugins,
         shared::{
@@ -716,6 +717,7 @@ pub mod prelude {
                 registry::rule_fns::RuleFns,
                 rules::{AppRuleExt, component::ReplicationMode},
                 signature::Signature,
+                visibility::{ComponentScope, FilterScope, SingleComponent, VisibilityFilter},
             },
             replicon_tick::RepliconTick,
         },
@@ -727,14 +729,9 @@ pub mod prelude {
     };
 
     #[cfg(feature = "server")]
-    #[expect(deprecated, reason = "Re-export of deprecated aliases")]
     pub use super::server::{
-        AuthorizedClient, PriorityMap, ServerPlugin, ServerSystems,
-        message::ServerMessagePlugin,
-        related_entities::SyncRelatedAppExt,
-        visibility::{
-            AppVisibilityExt, ComponentScope, FilterScope, SingleComponent, VisibilityFilter,
-        },
+        AuthorizedClient, PriorityMap, ServerPlugin, ServerSystems, message::ServerMessagePlugin,
+        related_entities::SyncRelatedAppExt, visibility::AppVisibilityExt,
     };
 
     #[cfg(feature = "client_diagnostics")]